调整内联编辑器滚动条
https://github.com/codebdy/rxdrag/issues/214
