+++ packages/rxdrag/editor/react-antd-shell-inline/src/components/Panel/index.tsx
@@ -5,7 +5,7 @@ import { boxShadow } from "../utils";
 export const Panel = styled(DraggableWidget).attrs({ resizable: true, maxWidth: 600, minWidth: 100, minHeight: 100 })`
   position: fixed;
   height:calc(100vh - 120px);
-  width: 200px;
+  width: 260px;
   box-shadow: ${boxShadow};
   top: 50%;
   transform: translateY(-50%);
