+++ packages/rxdrag/editor/react-antd-shell-inline/src/components/Toolkits/index.tsx
@@ -1,50 +1,21 @@
-import { memo, useMemo } from "react"
-import styled, { ThemeProvider } from "styled-components"
-import { Navbar } from "../Navbar"
-import { Toolbar } from "../Toolbar"
-import { Toolbox } from "../Toolbox"
-import { PropertyPanel } from "../PropertyPanel"
-import { GlobalToken, theme } from "antd"
+import { memo } from "react"
 import { ToolketsScope } from "./ToolketsScope"
+import { ToolkitsInner, ToolkitsInnerProps } from "./ToolkitsInner"
+import { ConfigProvider, theme } from "antd";
+import "./style.css"
 
-const Container = styled.div`
-  z-index: 100000;
-`
-
-export const Toolkits = memo((props: {
-  toolbox?: React.ReactNode,
-  toolbar?: React.ReactNode | false,
-  //用于localestorage缓存数据
+export const Toolkits = memo((props: ToolkitsInnerProps & {
   name?: string,
 }) => {
-  const { toolbox, toolbar, name } = props;
-  const { token } = theme.useToken()
-  const themeValue: { token: GlobalToken } = useMemo(() => {
-    return {
-      token
-    }
-  }, [token])
-  return (
-    <ToolketsScope name={name}>
-      <ThemeProvider theme={themeValue}>
-        <Container>
-          {
-            toolbox !== false &&
-            <Toolbox>
-              {toolbox}
-            </Toolbox>
-          }
+  const { toolbox, toolbar, name, themeMode = "dark" } = props;
 
-          <PropertyPanel />
-          {
-            toolbar !== false &&
-            <Toolbar>
-              {toolbar}
-            </Toolbar>
-          }
-          <Navbar />
-        </Container>
-      </ThemeProvider>
-    </ToolketsScope>
+  return (
+    <ConfigProvider
+      theme={{ algorithm: themeMode === "dark" ? theme.darkAlgorithm : theme.defaultAlgorithm }}
+    >
+      <ToolketsScope name={name}>
+        <ToolkitsInner toolbox={toolbox} toolbar={toolbar} themeMode={themeMode} />
+      </ToolketsScope>
+    </ConfigProvider>
   )
 })
