+++ packages/rxdrag/editor/react-antd-shell-inline/src/components/Toolkits/style.css
@@ -1,0 +1,21 @@
+.inline-editor ::-webkit-scrollbar {
+  display: block;
+  width: 0.35rem;
+  height: 0.35rem;
+}
+
+.inline-editor ::-webkit-scrollbar-track {
+  border-radius: 0;
+}
+
+.inline-editor ::-webkit-scrollbar-thumb {
+  border-radius: 0.2rem;
+  background: rgba(0, 0, 0, 0.3);
+  transition: all .2s;
+}
+
+.inline-editor.dark ::-webkit-scrollbar-thumb {
+  border-radius: 0.2rem;
+  background: rgba(255, 255, 255, 0.3);
+  transition: all .2s;
+}
